Refactor dead letter queue message functionality
